Add DM monitoring feature
Monitoring templates can be dropped in the Roboconf configuration directory:
- either in the monitoring-template directory, in this case the template is global to all applications.
- or in the {applicationName}/monitoring-template directory, in this case the template is specific to that application.

The templates use a {{mustache}} based syntax, extended with Handlebars features + Roboconf specific operators (helpers).
See the Roboconf Specification Request 232 (Monitoring Template Language) for a complete description of the template syntax.
The monitoring reports output syntax can be virtually anything (text-based).

The monitoring manager scans the Roboconf configuration directory for changes in the templates. If a template is added/modified,
the corresponding monitoring reports is generated/updated. Reports are *not* removed whan a template has gone.

The monitoring part is *optional*: it means the Roboconf DM can work perfectly without the presence of the monitoring manager bundle.